+++ Changelog.md
@@ -45,6 +45,7 @@
 - Fix: Module dupe involving doors
 - Fix: It's possible to connect one's own Password-protected Chest to that of another owner
 - Fix: Players hear a constant fire extinguishing sound when being in fake lava while it's raining
+- Fix: Alarm sound plays from the player's position instead of the alarm's position
 - Removed: Some unnecessary menu types. This may result in a "missing registry entries" message showing up, which can be accepted
 - Removed: "allowCodebreakerItem" config option. Disabling the Codebreaker is now achieveable by setting "codebreaker_chance" to a negative value
 - Misc.: More texture and model changes